implemen GATES in template blade

+++ resources/views/article.blade.php
@@ -31,17 +31,21 @@
 
                     </div>
                     <div class="d-flex justify-content-end mt-3">
-                        <a href="article/edit/{{ $articles->id }}" class="btn btn-sm me-2 border" style="background: transparent; color: #000;">
-                            <i class="bi bi-pencil-square"></i> Edit
-                        </a>
-                        <a href="article/delete/{{ $articles->id }}" class="btn btn-sm btn-danger">
-                            <i class="bi bi-trash"></i> Delete
-                        </a>
+                        @can('update', $articles)
+                            <a href="article/edit/{{ $articles->id }}" class="btn btn-sm me-2 border" style="background: transparent; color: #000;">
+                                <i class="bi bi-pencil-square"></i> Edit
+                            </a>
+                        @endcan
+                        @can('delete', $articles)
+                            <a href="article/delete/{{ $articles->id }}" class="btn btn-sm btn-danger">
+                                <i class="bi bi-trash"></i> Delete
+                            </a>
+                        @endcan
                     </div>
                     </div>
-    
-                </div> 
-             
+
+                </div>
+
     </a>
 @endforeach
 </div>
